Create route53_reusable_delegation_set_finder.py
create a reusable_delegation_set that contains a name server you want  by just brute force create and delete the reusable_delegation_sets  in a loop.

 if you get one you want you can use it via "aws route53 create-hosted-zone --name demo.example.com --caller-reference 202x-Mon-day-<random> --delegation-set-id $delegation_set_id

TODO: update script so it just tries to create the hosted domain using the delegation set. 

this does not seem to work for sitting duck lame delegations, but more research is needed.